feat: add copy fallback
